fix bot impersonating others

+++ main.py
@@ -15,17 +15,28 @@ class EchoBot(discord.Bot):
 
 bot = EchoBot(intents=discord.Intents(messages=True))
 
+def extract_id(text: str) -> int | None:
+  try:
+    return int(text.strip('<@').strip('>'))
+  except ValueError:
+    return None
+
 @bot.event
 async def on_message(message: discord.Message):
   if message.author == bot.user: return
+  # bot should not impersonate others
+  if message.author.bot: return
   if len(message.content) <= 0: return
   if len(message.mentions) <= 0: return
-  if message.mentions[0].id != bot.user.id: return
-  await message.delete()
   content = message.content.split(' ', 1)
   if len(content) == 1:
     return await message.channel.send('找我嗎？')
-  text = message.content.split(' ', 1)[1]
+  # check if the first mention is actually the bot (this is for replying with mention)
+  actual_id = extract_id(content[0])
+  if not actual_id: return
+  if actual_id and actual_id != bot.user.id: return
+  if not message.author.bot: await message.delete()
+  text = content[1]
   if message.reference and message.reference.resolved and isinstance(message.reference.resolved, discord.Message):
     await message.reference.resolved.reply(text, mention_author=False)
   else:
